+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.6.0
+
+- add `publicDirectory`, `buildDirectory`, `refresh`, `verbose` plugin option
+- add `dev-server-404.html` page
+
 ## v0.5.2
 
-add `servePublic` plugin option
+- add `servePublic` plugin option